requirements.txt
absl-py==1.0.0
alembic==1.7.7
asgiref==3.3.4
astunparse==1.6.3
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cryptography==37.0.2
cycler==0.11.0
decouple==0.0.7
distlib==0.3.4
dj-database-url==0.5.0
Django==3.2
django-heroku==0.3.1
django-jazzmin==2.4.9
django-multiforloop==0.3.1
django-widget-tweaks==1.4.12
filelock==3.6.0
Flask==2.1.1
Flask-Dropzone==1.6.0
Flask-Migrate==3.1.0
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
flatbuffers==2.0
fonttools==4.31.2
gast==0.5.3
google-auth==2.6.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.44.0
gunicorn==20.1.0
h5py==3.6.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==13.0.0
Mako==1.2.0
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
mysql-connector-python==8.0.28
mysqlclient==2.1.0
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
Pillow==9.0.1
platformdirs==2.5.1
protobuf==3.20.0
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
sorl-thumbnail==12.8.0
SQLAlchemy==1.4.36
sqlparse==0.4.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
twilio==7.9.1
typing-extensions==3.7.4.3
urllib3==1.26.9
virtualenv==20.14.0
Werkzeug==2.1.1
whitenoise==6.0.0
wrapt==1.14.0
WTForms==3.0.1
